// client/src/enums/product-action.ts
export enum ProductAction {
    PRODUCT_REQUEST_START = 'PRODUCT_REQUEST_START',
    PRODUCT_REQUEST_FAILURE = 'PRODUCT_REQUEST_FAILURE',
    GET_PRODUCTS_SUCCESS = 'GET_PRODUCTS_SUCCESS',
    ADD_PRODUCT_SUCCESS = 'ADD_PRODUCT_SUCCESS',
    REMOVE_PRODUCT_SUCCESS = 'REMOVE_PRODUCT_SUCCESS',
    UPDATE_PRODUCT_SUCCESS = 'UPDATE_PRODUCT_SUCCESS',
}